motify_back/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.2
sqlalchemy==2.0.19
alembic==1.11.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
asyncpg
pydantic[email]==2.0.3
python-dotenv==1.0.0
pydantic-settings==2.0.3
python-multipart
pillow
redis==5.0.1
psycopg2-binary==2.9.9

# SQLAlchemy[mypy] si usas mypy para type hints más estrictos con SQLAlchemy
# python-multipart para subida de archivos si lo necesitas más adelante
